requirements.txt
alembic==1.7.5
boto3==1.20.7
botocore==1.23.7
Cerberus==1.3.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
decorator==5.1.0
Flask==2.0.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
flask-crontab==0.1.2
Flask-Migrate==3.1.0
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.1.5
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
numpy==1.21.4
Pillow==8.4.0
plum-py==0.8.0
protobuf==3.19.1
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-resize-image==1.1.20
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.27
typing-extensions==4.1.1
urllib3==1.26.7
validate-docbr==1.10.0
validators==0.18.2
Werkzeug==2.0.2
zipp==3.7.0
mysqlclient==2.0